separated out the events list and renamed the lists and sections
